addSkill in skill.service ausgelagert.
