+++ src/components/skillformular.tsx
@@ -1,8 +1,7 @@
 import React, { useState, ChangeEvent } from "react";
 import { Link } from "react-router-dom";
-import { v4 as uuidv4 } from "uuid";
-import { Skill } from "../types/skill"; // Import the Skill interface
 import { skillsPath } from "../app-paths";
+import {addSkill} from "../services/skill.service";
 
 const SkillForm: React.FC = () => {
   const [skillname, setSkillname] = useState<string>("");
@@ -14,31 +13,14 @@ const SkillForm: React.FC = () => {
   };
 
   const handleCreateSkill = () => {
-    if (skillname.trim().length === 0) {
+
+    if (skillname === null) {
       setValidationMessage("Bitte fülle das Feld aus");
-      return;
-    }
-
-    const newSkill: Skill = {
-      id: uuidv4(),
-      name: skillname,
-      level: 0,
-      created: new Date(),
-    };
-
-    const localSkills: string | null = window.localStorage.getItem("skills");
-    let updatedSkillList: Skill[] = [];
-
-    if (localSkills !== null) {
-      const existingSkills: Skill[] = JSON.parse(localSkills);
-      updatedSkillList = [...existingSkills, newSkill];
     } else {
-      updatedSkillList = [newSkill];
+      setSkillname("");
+      setValidationMessage("");
+      addSkill(skillname);
     }
-
-    window.localStorage.setItem("skills", JSON.stringify(updatedSkillList));
-    setSkillname("");
-    setValidationMessage("");
   };
 
   return (
