+++ src/types/skill.ts
@@ -1,4 +1,4 @@
-export interface Skill  {
+export type Skill = {
   id: string;
   name: string;
   level: number;
